all grpc and rest routes

+++ currency/types/interfaces/rate_store.go
@@ -6,12 +6,12 @@ type RateStoreRow struct {
 	Id                  uint32    `json:"id"`
 	BaseCurrency        string    `json:"base_currency"`
 	DestinationCurrency string    `json:"destination_currency"`
-	Rate                float32   `json:"rate"`
+	Rate                float64   `json:"rate"`
 	UpdatedAt           time.Time `json:"updated_at"`
 }
 
 type RateStore interface {
 	GetRate(string, string) (*RateStoreRow, error)
-	Upsert(string, string, float32) error
+	Upsert(string, string, float64) error
 	Delete(string, string) error
 }
